webview: add metadata field for backend communication
There is now a separate argument called "metadata" which can be used to
attach metadata to be read by the backend on how to handle the called
function. e.g. how to handle the logging for this called function / if
the logs should be added to a specific group etc.